\name{fishsizespectrum-package}
\alias{fishsizespectrum-package}
\alias{fishsizespectrum}
\docType{package}
\title{
\packageTitle{fishsizespectrum}
}
\description{
\packageDescription{fishsizespectrum}
}
\details{
The most important functions are
\itemize{
\item baseparameters() -- returns a structure with all default parameters
\item spectrum(parameters) -- returns the population size spectrum given a list of parameters.
\item calcrefpoints(parameters) -- calculates all fisheries reference points
\item calcSelectionResponse(QGparameters,W,F) -- calculates selection responses
}
%The DESCRIPTION file:
%\packageDESCRIPTION{fishsizespectrum}
%\packageIndices{fishsizespectrum}
%~~ An overview of how to use the package, including the most important ~~
%~~ functions ~~
}
\author{
\packageAuthor{fishsizespectrum}

Maintainer: \packageMaintainer{fishsizespectrum}
}

\references{
The details of the mathematical models and all parameters are in: K.H. Andersen (2019): Fish, Ecology, Evolution, and Exploitation. Princeton University Press.
}

\keyword{ package }

\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}

